app.py
############################################################
# FACE ATTENDANCE SYSTEM – FINAL VERSION
# Core Model: InsightFace buffalo_l
# Strategy: AI recognition + Manual verification
# Result: 100% final attendance correctness
############################################################

import os
import cv2
import numpy as np
import pandas as pd
import streamlit as st
from datetime import datetime, date
from PIL import Image, ImageDraw, ImageFont
from insightface.app import FaceAnalysis

############################################################
# CONFIG
############################################################
st.set_page_config(page_title="Face Attendance (buffalo_l)", layout="wide")

ROOT_REG = "registered_faces"
ROOT_SAVE = "attendance_records"

os.makedirs(ROOT_REG, exist_ok=True)
os.makedirs(ROOT_SAVE, exist_ok=True)

# 🔥 buffalo_l plays the MAIN role here
MODEL_NAME = "buffalo_l"

THRESHOLD = 0.46          # strict threshold for high precision
MIN_FACE_RATIO = 0.012   # ignore very small faces

############################################################
# LOAD buffalo_l MODEL
############################################################
@st.cache_resource
def load_model():
    model = FaceAnalysis(name=MODEL_NAME)
    model.prepare(ctx_id=-1, det_size=(640, 640))  # CPU
    model.get(np.zeros((320, 320, 3), dtype=np.uint8))  # warm-up
    return model

model = load_model()

############################################################
# LOAD REGISTERED STUDENT FACES
############################################################
@st.cache_data
def load_registered_faces():
    names, embs = [], []

    for fn in os.listdir(ROOT_REG):
        path = os.path.join(ROOT_REG, fn)
        img = cv2.imread(path)
        if img is None:
            continue

        faces = model.get(img)
        if not faces:
            continue

        emb = faces[0].normed_embedding.astype(np.float32)
        emb /= np.linalg.norm(emb) + 1e-10

        base = os.path.splitext(fn)[0]
        if "_" in base and base.split("_")[-1].isdigit():
            student_id = "_".join(base.split("_")[:-1])
        else:
            student_id = base

        names.append(student_id)
        embs.append(emb)

    if not embs:
        return [], np.zeros((0, 512), dtype=np.float32)

    return names, np.vstack(embs)

names, reg_embs = load_registered_faces()

############################################################
# HELPERS
############################################################
def pil_from_bgr(img):
    return Image.fromarray(cv2.cvtColor(img, cv2.COLOR_BGR2RGB))

def draw_box(draw, bbox, label, green=True):
    x1, y1, x2, y2 = map(int, bbox)
    color = (0, 200, 0) if green else (220, 50, 50)
    draw.rectangle([x1, y1, x2, y2], outline=color, width=3)

    try:
        font = ImageFont.truetype("DejaVuSans.ttf", 14)
    except:
        font = ImageFont.load_default()

    draw.text((x1, max(0, y1 - 18)), label, fill=color, font=font)

############################################################
# SIDEBAR
############################################################
with st.sidebar:
    st.header("Session Details")
    subject = st.text_input("Subject")
    faculty = st.text_input("Faculty")
    period = st.selectbox("Period", [f"Period {i}" for i in range(1, 9)])

    st.markdown("---")
    st.write(f"Registered students: **{len(set(names))}**")

    if st.button("Reload Registry"):
        load_registered_faces.clear()
        st.rerun()

if not names:
    st.warning("No registered faces found in registered_faces/")
    st.stop()

############################################################
# MAIN UI
############################################################
st.title("📸 Face Attendance System (buffalo_l + Manual)")

uploaded = st.file_uploader(
    "Upload ONE classroom group photo",
    type=["jpg", "jpeg", "png"]
)

if uploaded:
    img = cv2.imdecode(np.frombuffer(uploaded.read(), np.uint8), cv2.IMREAD_COLOR)
    faces = model.get(img)

    pil_img = pil_from_bgr(img)
    draw = ImageDraw.Draw(pil_img)

    recognised = set()

    for face in faces:
        x1, y1, x2, y2 = map(int, face.bbox)
        face_area = (x2 - x1) * (y2 - y1)
        img_area = img.shape[0] * img.shape[1]

        if face_area / img_area < MIN_FACE_RATIO:
            draw_box(draw, face.bbox, "Ignored", False)
            continue

        emb = face.normed_embedding.astype(np.float32)
        emb /= np.linalg.norm(emb) + 1e-10

        sims = reg_embs @ emb
        idx = int(np.argmax(sims))
        score = float(sims[idx])

        if score >= THRESHOLD:
            name = names[idx]
            recognised.add(name)
            draw_box(draw, face.bbox, name, True)
        else:
            draw_box(draw, face.bbox, "Unknown", False)

    st.image(pil_img, use_container_width=True)

    ########################################################
    # ATTENDANCE TABLE
    ########################################################
    st.subheader("📋 Attendance")

    all_students = sorted(set(names))
    attendance = []

    for s in all_students:
        attendance.append({
            "Student Name": s.replace("_", " "),
            "Status": "Present" if s in recognised else "Absent"
        })

    df = pd.DataFrame(attendance)
    st.dataframe(df, hide_index=True)

    ########################################################
    # MANUAL VERIFICATION
    ########################################################
    st.subheader("🟥 Manual Verification (Red Faces)")

    for s in all_students:
        if s not in recognised:
            if st.checkbox(f"Mark {s.replace('_', ' ')} as Present"):
                df.loc[df["Student Name"] == s.replace("_", " "), "Status"] = "Present"

    ########################################################
    # SAVE ATTENDANCE
    ########################################################
    if st.button("💾 Save Attendance"):
        today = date.today().isoformat()
        folder = os.path.join(ROOT_SAVE, today)
        os.makedirs(folder, exist_ok=True)

        df["Subject"] = subject
        df["Faculty"] = faculty
        df["Period"] = period
        df["Date"] = today
        df["Time"] = datetime.now().strftime("%I:%M %p")

        file_path = os.path.join(folder, f"{today}_{period.replace(' ', '_')}.csv")
        df.to_csv(file_path, index=False)

        st.success("✅ Attendance saved successfully")
