requirements.txt
streamlit
insightface==0.7.3
onnxruntime>=1.20,<1.22
opencv-python
pillow
numpy
pandas
